requirements.txt
eyed3
python-dotenv
